Put personactions into a toolbar
Implement the idea from #4.
This frees up some space that we can use to display more information about the contact in the future.
The call/sms/email actions are hardcoded instead of coming from the actions plugin. This will simplify the implementation for multiple phone numbers (which is not done in this patch). The IMPP action still comes from the plugin since that's a bit more complex.
If there are more actions than horizontal space they are added into an overflow menu automatically. There is a minor bug in Kirigami that shows an empty menu sometimes (https://bugs.kde.org/show_bug.cgi?id=415412), but that shouldn't be a blocker for this.