Remove call action from contacts list
Scrolling the contact list feels a bit sluggish, pontentially because of
rather heavy delegates.

The inline call action also does not provide much value since there is a
call action on the details page.

It also does not handle multiple phonenumbers.

For quickly calling a number the dialer app is the better choice